+++ core/dbt/artifacts/resources/__init__.py
@@ -6,6 +6,11 @@ from dbt.artifacts.resources.v1.macro import Macro, MacroDependsOn, MacroArgumen
 from dbt.artifacts.resources.v1.docs import Docs
 from dbt.artifacts.resources.v1.group import Group
 from dbt.artifacts.resources.v1.owner import Owner
+from dbt.artifacts.resources.v1.saved_query import (
+    Export,
+    ExportConfig,
+    QueryParams,
+)
 from dbt.artifacts.resources.v1.semantic_layer_components import (
     FileSlice,
     SourceFileMetadata,
